監視するパラメータの変更、スクリプトの移動、メニュー項目を４つにしてTMPをtextに変更、Canvasのプレハブ化
