+++ Assets/Scripts/GUI/PlayerGuage.cs
