+++ Assets/Scripts/GUI/PlayerGuage.cs.meta
